+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+* [PR-86](https://github.com/ITK-Leantime/project-overview/pull/86)
+  * Fix empty status dropdowns due to array_merge renumbering project-ID keys.
+  * Fix issue with length of status title exceeding dropdown max width.
+
 ## [4.2.0] - 2026-05-20
 
 * [PR-83](https://github.com/ITK-Leantime/project-overview/pull/83)